Move _no_chain_background_task to reflex.event
It's right next to the definition of rx.background itself and only defined
once, then used within the State.
